Spell effects
Dot damage and enemy slow down. experimented with a damage popup to show how much damage is done to an enemy. Everything works except it spawns on top of the player instead of on top of the enemy and no matter what i try it will not stop doing that.